Create from 4l template, copy in existing design
Intent is to test the ESP32-S3 and ESP32-C3 bare chips (with qio flash
IC and clock chip connected). If it works, it would greatly reduce the
price of small-batch boards (because ESP32 modules cannot be placed in
JLCPCB's economic assembly).

Was already working on a design using an ESP32-S3 bare chip, so have
copied int ehr elevant parts of that.