Add sheet for ESP32-C3 bare chip as well.
Similar to the ESP32-S3 circuit, so I copied most of it (one capacitor
is not needed, according to the reference schematics).

No flash chip is attached; makes sense to use one of the embedded flash
chips to test with, as it makes for a very small design (and 16MB is
really not needed).