Add 5v and 3v3 buck regulators, brige rectifier
Buck regulators are copied from existing designs. 5V regulator needs to
be reworked though; due to ESD vulnerability.

Added 4 diodes in bridge rectifier configuration. Provides a simpler
solution than the polarity-protected eFuse. Due to how the
buck-regulators work, it should be fine for any voltage from ~4.8V 17V
(max Vin of buck regulator). Downside is icnreased power-dissipation,
but since this board is not intended to use a lot of current it should
be good (the USB port can bypass this too, but only for 5V)