Apply XOR to 5V buck reg and USB power
Some 5V regulators will be very unhappy if a higher-voltage is applied -
e.g. from USB. Therefore, use a comparator to turn off the load-switch
when 12V power is present.

Circuit got a bit more complex because it needs to deal with dynamic
situations as things are plugged in. Will still need some work, likely.