Replace buck-regulator with buck-boost regulator
Replaced the buck-regulator with a buck-boost regulator. The primary
purpose is because the buck-regulator could fail when a higher-voltage
output was conected. The buck-regulators internal switches and design
should be much less vulnerable for this.

A second benefit of the buck-boost regulator is thata 5V input can now
still be used, as it accounts for the voltage-drop from the
bridge-rectifier diodes.